+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## Unreleased
+
+## 2.78.0
+* Add support for `OAUTH_ACCESS_REVOKED` in `WebhookNotification`s
+* Add support for dispute search by `customerId`, `disbursementDate`, and `effectiveDate`
+* Add `payerId` accessor in `PayPalAccount`
+* Add support for VCR compelling evidence dispute representment
+
 ## 2.77.0
 * Fix possible NullPointerException in UsBankAccount
 * Add support for `association_filter_id` in `Customer#find`